requirements.txt
streamlit
groq
python-dotenv
sounddevice
scipy
pydub
streamlit-audiorecorder